+++ engine/audio/private/snd_wave_data.h
@@ -29,7 +29,7 @@ public:
 abstract_class IWaveStreamSource
 {
 public:
-	virtual int		UpdateLoopingSamplePosition( int samplePosition ) = 0;
+	virtual int		UpdateLoopingSamplePosition( int64 samplePosition ) = 0;
 	virtual void	UpdateSamples( char *pData, int sampleCount ) = 0;
 	virtual int		GetLoopingInfo( int *pLoopBlock, int *pNumLeadingSamples, int *pNumTrailingSamples ) = 0;
 };
